pyproject.toml
[project]
name = "api-secret-word"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.1.0",
    "tortoise-orm[asyncpg]>=0.25.0",
    "uvicorn>=0.34.2",
]
